Removed old Tooltip component
