+++ web/src/components/tooltip.js
@@ -1,83 +1,54 @@
-/* eslint-disable */
-// TODO: remove once refactored
+import React, { useRef } from 'react';
+import { connect } from 'react-redux';
 
-import { event as d3Event } from 'd3-selection';
+import { useWidthObserver, useHeightObserver } from '../effects';
 
-const d3 = Object.assign(
-  {},
-  require('d3-selection'),
-  require('d3-transition'),
-);
+const MARGIN = 16;
 
-function placeTooltip(selector, eventX, eventY) {
-  const tooltip = d3.select(selector);
-  const w = tooltip.node().getBoundingClientRect().width;
-  const h = tooltip.node().getBoundingClientRect().height;
-  const margin = 16;
+const mapStateToProps = state => ({
+  position: state.application.tooltipPosition,
+});
+
+const Tooltip = ({ id, children, position }) => {
+  const ref = useRef(null);
+
+  const width = useWidthObserver(ref);
+  const height = useHeightObserver(ref);
+
   const screenWidth = window.innerWidth;
   const screenHeight = window.innerHeight;
 
+  if (!position) return null;
+
+  const style = {};
   let x = 0;
   let y = 0;
   // Check that tooltip is not larger than screen
   // and that it does fit on one of the sides
-  if (2 * margin + w >= screenWidth ||
-    (eventX + w + margin >= screenWidth && eventX - w - margin <= 0 )) {
+  if (2 * MARGIN + width >= screenWidth
+    || (position.x + width + MARGIN >= screenWidth && position.x - width - MARGIN <= 0)) {
     // TODO(olc): Once the tooltip has taken 100% width, it's width will always be 100%
     // as we base our decision to revert based on the current width
-    tooltip.style('width', '100%');
+    style.width = '100%';
   } else {
-    x = eventX + margin;
+    x = position.x + MARGIN;
     // Check that tooltip does not go over the right bound
-    if (w + x >= screenWidth) {
+    if (width + x >= screenWidth) {
       // Put it on the left side
-      x = eventX - w - margin;
+      x = position.x - width - MARGIN;
     }
   }
-  y = eventY - margin - h;
-  if (y < 0) y = eventY + margin;
-  if (y + h + margin >= screenHeight) y = eventY - h - margin;
-  // y = eventY
-  tooltip
-    .style('transform', `translate(${x}px,${y}px)`);
-}
+  y = position.y - MARGIN - height;
+  if (y < 0) y = position.y + MARGIN;
+  if (y + height + MARGIN >= screenHeight) y = position.y - height - MARGIN;
 
-function Tooltip(selector) {
-    this._selector = selector
-    var that = this;
-    d3.select(this._selector)
-        .style('opacity', 0)
-        // For mobile, hide when tapped
-        .on('click', function(e) { that.hide(); d3Event.stopPropagation(); })
-    return this;
-}
+  style.transform = `translate(${x}px,${y}px)`;
 
-Tooltip.prototype.show = function() {
-    if (this.isShowing) { return; }
-    this.isShowing = true;
-    this.isVisible = true;
-    d3.select(this._selector)
-        .style('display', 'block')
-        .transition()
-        .style('opacity', 1)
-        .on('end', () => { this.isShowing = false; });
-    return this;
-}
-Tooltip.prototype.update = function(x, y) {
-    placeTooltip(this._selector, x, y);
-    return this;
-}
-Tooltip.prototype.hide = function() {
-    this.isShowing = false;
-    this.isVisible = false;
-    d3.select(this._selector)
-        .style('width', null) // this is a temporary fix for the 100% width problem
-        .transition()
-        .style('opacity', 0)
-        .on('end', function() {
-            d3.select(this).style('display', 'none');
-        });
-    return this;
-}
+  return (
+    <div id={id} className="tooltip panel" style={style} ref={ref}>
+      {children}
+    </div>
+  );
+};
 
-export default Tooltip
+export default connect(mapStateToProps)(Tooltip);
