+++ web/src/components/tooltips/lowcarboninfotooltip.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import { __ } from '../../helpers/translation';
 import { LOW_CARBON_INFO_TOOLTIP_KEY } from '../../helpers/constants';
 
-import TooltipContainer from './tooltipcontainer';
+import Tooltip from '../tooltip';
 
 const mapStateToProps = state => ({
   visible: state.application.tooltipDisplayMode === LOW_CARBON_INFO_TOOLTIP_KEY,
@@ -14,12 +14,12 @@ const LowCarbonInfoTooltip = ({ visible }) => {
   if (!visible) return null;
 
   return (
-    <TooltipContainer id="lowcarb-info-tooltip">
+    <Tooltip id="lowcarb-info-tooltip">
       <b><span id="lowcarb-info-title">{__('tooltips.lowcarbon')}</span></b>
       <br />
       <small><span id="lowcarb-info-text">{__('tooltips.lowCarbDescription')}</span></small>
       <br />
-    </TooltipContainer>
+    </Tooltip>
   );
 };
